+++ TwoPaneLayout/library/src/main/java/com/microsoft/device/dualscreen/twopanelayout/TwoPaneMode.kt
@@ -1,0 +1,13 @@
+package com.microsoft.device.dualscreen.twopanelayout
+
+/**
+ * TwoPaneMode
+ * TwoPane,          always shows two panes, regardless the orientation, by default
+ * HorizontalSingle  shows big single pane in horizontal orientation layout(top/bottom)
+ * VerticalSingle    shows big single pane in vertical orientation layout(left/right)
+ */
+enum class TwoPaneMode {
+    TwoPane,
+    HorizontalSingle,
+    VerticalSingle
+}
