+++ TwoPaneLayout/library/src/main/java/com/microsoft/device/dualscreen/twopanelayout/TwoPaneScope.kt
@@ -1,0 +1,33 @@
+package com.microsoft.device.dualscreen.twopanelayout
+
+import androidx.compose.foundation.layout.LayoutScopeMarker
+import androidx.compose.runtime.Immutable
+import androidx.compose.runtime.Stable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.debugInspectorInfo
+
+@LayoutScopeMarker
+@Immutable
+interface TwoPaneScope {
+    @Stable
+    fun Modifier.weight(
+        weight: Float,
+    ): Modifier
+}
+
+internal object TwoPaneScopeInstance : TwoPaneScope {
+    @Stable
+    override fun Modifier.weight(weight: Float): Modifier {
+        require(weight > 0.0) { "invalid weight $weight; must be greater than zero" }
+        return this.then(
+            LayoutWeightImpl(
+                weight = weight,
+                inspectorInfo = debugInspectorInfo {
+                    name = "weight"
+                    value = weight
+                    properties["weight"] = weight
+                }
+            )
+        )
+    }
+}
